[REFACTOR/#310] mapping 패키지 파일 구조 변경
