+++ src/main/java/com/assu/server/domain/admin/dto/StudentAdminRequestDTO.java
@@ -1,0 +1,5 @@
+package com.assu.server.domain.admin.dto;
+
+public class StudentAdminRequestDTO {
+
+}
