+++ src/main/java/com/assu/server/domain/admin/entity/StudentAdmin.java
@@ -1,8 +1,7 @@
-package com.assu.server.domain.mapping.entity;
+package com.assu.server.domain.admin.entity;
 
 import com.assu.server.domain.common.entity.BaseEntity;
 import com.assu.server.domain.user.entity.Student;
-import com.assu.server.domain.admin.entity.Admin;
 import jakarta.persistence.*;
 import lombok.*;
