--- src/main/java/com/assu/server/domain/mapping/dto/StudentAdminRequestDTO.java
@@ -1,5 +1,0 @@
-package com.assu.server.domain.mapping.dto;
-
-public class StudentAdminRequestDTO {
-
-}
